artifact_restoration/requirements.txt
# Artifact Restoration Multi-Agent System
# Requirements

# Core dependencies
google-generativeai>=0.3.0
flask>=3.0.0
Pillow>=10.0.0

# Additional utilities
python-dotenv>=1.0.0
werkzeug>=3.0.0
